Split Self expression validation
